abstract out the uplot graph height handler
